+++ CHANGELOG.md
@@ -1,22 +1,37 @@
 # CHANGELOG
 
-## 1.3.1
+## 1.4.0 - 2019-04-25
+
+### Added
+
+- Add HtmlSpecialChars::x() method
+- Add HtmlSpecialCharsArrays::x() method
+
+## 1.3.1 - 2019-04-17
+
+### Fixed
 
 - Fix bug in PathSegment::x() when $url parameter has no segments
 
-## 1.3.0
+## 1.3.0 - 2019-04-16
+
+### Added
 
 - Add AnchorTag::x() method
 - Add ImageTag::x() method
 
-## 1.2.0
+## 1.2.0 - 2019-04-07
+
+### Added
 
 - Add NullifiedArray::x() method
 
-## 1.1.0
+## 1.1.0 - 2019-04-05
+
+### Added
 
 - Add Redirect::x() method
 
-## 1.0.0
+## 1.0.0 - 2019-03-20
 
 - Release first version